GERONIMO-5567 clean up unused classes and code
git-svn-id: https://svn.apache.org/repos/asf/geronimo/server/trunk@995315 13f79535-47bb-0310-9956-ffa450edef68

+++ plugins/jetty8/geronimo-jetty8/src/main/java/org/apache/geronimo/jetty8/JettyContainerImpl.java
@@ -206,7 +206,7 @@ public class JettyContainerImpl implements JettyContainer, SoapHandler, GBeanLif
             stats.getResponses5xxImpl().setCount((long) statsHandler.getResponses5xx());
 
             /* set elapsed time for stats collection */
-            stats.getStatsOnMsImpl().setCount((long) statsHandler.getStatsOnMs());
+            stats.getStatsOnMsImpl().setCount(statsHandler.getStatsOnMs());
         }
         return stats;
     }
@@ -288,6 +288,7 @@ public class JettyContainerImpl implements JettyContainer, SoapHandler, GBeanLif
         try {
             server.stop();
         } catch (Exception e) {
+            //ignore
         }
     }
 
@@ -295,7 +296,7 @@ public class JettyContainerImpl implements JettyContainer, SoapHandler, GBeanLif
         try {
             server.stop();
         } catch (Exception e) {
-            // continue
+            //ignore
         }
     }
